+++ src/indexHelpers.ts
@@ -1,4 +1,5 @@
 export { default as findScrollParent } from "./helpers/findScrollParent";
+export { findScrollParentAll } from "./helpers/findScrollParent";
 export { default as focusFirst } from "./helpers/focusFirst";
 export { default as nestedProperty } from "./helpers/nestedProperty";
 export { default as observer } from "./helpers/observer";