+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 **Features**:
 
 - Optimized for webpack (via [sideEffects](https://webpack.js.org/guides/tree-shaking/#mark-the-file-as-side-effect-free))
+- Added [helpers](README.md#helpers)
+  - `findScrollParentAll`
 - [`FormControls`](README.md#controls)
   - Added TextControl (desc. coming soon...)
   - Added SelectControl (desc. coming soon...)